feat: 각 모듈에 맞게 Place 에 imageUrl, placeUrl 속성 생성
