+++ data/src/main/java/com/wakeup/data/network/mapper/PlaceResponseMapper.kt
@@ -1,0 +1,22 @@
+package com.wakeup.data.network.mapper
+
+import com.wakeup.data.database.entity.LocationEntity
+import com.wakeup.data.database.mapper.toDomain
+import com.wakeup.data.network.response.PlaceResponseItem
+import com.wakeup.domain.model.Place
+
+
+fun PlaceResponseItem.toDomain(): Place {
+    return Place(
+        mainAddress = place_name,
+        detailAddress = road_address_name,
+        imageUrl = getImageUrl(place_url),
+        location = LocationEntity(y.toDouble(), x.toDouble()).toDomain()
+    )
+}
+
+fun getImageUrl(placeUrl: String): String {
+
+
+    return " "
+}
