+++ data/src/main/java/com/wakeup/data/network/response/PlaceResponse.kt
@@ -1,8 +1,5 @@
 package com.wakeup.data.network.response
 
-import com.wakeup.data.database.mapper.toDomain
-import com.wakeup.data.database.entity.LocationEntity
-import com.wakeup.domain.model.Place
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -29,17 +26,10 @@ data class PlaceResponse(
 data class PlaceResponseItem(
     val place_name: String,
     val road_address_name: String,
+    val place_url: String,
     val x: String,
     val y: String
 )
-
-fun PlaceResponseItem.toDomain(): Place {
-    return Place(
-        mainAddress = place_name,
-        detailAddress = road_address_name,
-        location = LocationEntity(y.toDouble(), x.toDouble()).toDomain()
-    )
-}
 
 
 //@Serializable
