+++ domain/src/main/java/com/wakeup/domain/model/Place.kt
@@ -3,5 +3,6 @@ package com.wakeup.domain.model
 data class Place(
     val mainAddress: String,
     val detailAddress: String,
-    val location: Location
+    val imageUrl: String? = null,
+    val location: Location,
 )
